Move password entry to Profile, I18N stuff
